Only operate on active outputs
There was a bug where colord-kde registered colord devices for all connected
outputs during startup, even those which were not active at that time. However,
accessing their gamma led to an empty result which was treated as an error later
on.

In practice, the error meant that once kded was started while LVDS was connected
but inactive and DP2 connected and active, when the user enabled LVDS later on,
the profile of that output could not be changed. This happened because the code
already registered an output with colord, so a re-registration failed at the
time the output was enabled, and therefore any attempt at working with the newly
enabled output accessed an invalid CRTC, leading to invalid gamma and therefore
no profile manipulation.

Many thanks to Pascal de Bruijn for handing me through colord and general color
management troubleshooting and to Daniel Nicoletti for his feedback on how to
fix this within colord-kde.