Add CMake option to build WITHOUT_X11
We want to be able to build without X11 support even if some of the used
libraries may not work w/o X11 themselves yet or need to be built with X11
support for other reverse dependencies.

X11_FOUND is set automagically so far, and using
-DCMAKE_DISABLE_FIND_PACKAGE_X11 will break if any dependencies list X11
as required in their cmake config.

Introducing this option means there is no behavior change by default,
cmake will just skip finding X11 if the option is enabled.

Signed-off-by: Andreas Sturmlechner <asturm@gentoo.org>